add .readthedocs.yml
update .travis.yml

update setup.py and tests

tweak copyright dates

fix name error

fix name of stats file

fix name of exposures file

working on exposures_surveysim

remove extname keywords

remove extname keywords

remove comment

update minitest script

fix filenames in DESI_TARGET

fix filenames in DESI_TARGET

remove QUICKLOOK

remove link to QUICKLOOK

working on simspec

working on simspec

skip TNULL keywords

extra column

working on mtl

working on mtl

working on sky

working on sky

working on sky

working on targets

working on targets

working on targets

working on targets

working on truth [skip ci]

working on tile [skip ci]

working on tile [skip ci]

working on tile [skip ci]

working on tile [skip ci]

update logic for extname in hdu 0

fix table border

working on zcatalog [skip ci]

working on sky [skip ci]

working on sky [skip ci]

working on calib [skip ci]

working on cframe [skip ci]

working on cframe [skip ci]

working on frame [skip ci]

working on spectra [skip ci]

working on spectra [skip ci]

working on spectra [skip ci]

working on zbest [skip ci]

working on zbest [skip ci]
