+++ LICENSE.rst
@@ -1,4 +1,4 @@
-Copyright (c) 2014-2017, DESI Collaboration <desi-data@desi.lbl.gov>
+Copyright (c) 2014-2019, DESI Collaboration <desi-data@desi.lbl.gov>
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
