+++ doc/DESISURVEY_OUTPUT/exposures.rst
@@ -3,8 +3,8 @@ exposures
 =========
 
 :Summary: Record of simulated exposures.
-:Naming Convention: ``exposures.fits``
-:Regex: ``exposures\.fits``
+:Naming Convention: ``exposures_surveysim.fits``
+:Regex: ``exposures_surveysim\.fits``
 :File Type: FITS, 2 MB  (scales with the number of exposures)
 
 *Note*: currently this is only an output from surveysim, but it may become an
@@ -40,9 +40,7 @@ KEY     Example Value     Type Comment
 ======= ================= ==== =======
 TILES   desi--tiles.fits  str  Name of the tiles file specified in desisurvey config.
 NEXP    41875             int  Number of exposures recorded so far.
-COMMENT Baseline (seed=1) str  Comment describing this simulation.
 INITIAL 2019-12-01        str  YEAR-MM-DD of initial night used for integer offsets (or blank before any observing).
-EXTNAME META              str  Extension name.
 ======= ================= ==== =======
 
 Empty HDU.
@@ -62,7 +60,6 @@ KEY     Example Value Type Comment
 ======= ============= ==== =====================
 NAXIS1  41875         int  length of dimension 1
 NAXIS2  8             int  length of dimension 2
-EXTNAME EXPOSURES     str  extension name
 ======= ============= ==== =====================
 
 Required Data Table Columns
@@ -76,7 +73,7 @@ EXPTIME  float32 s      Length of time shutter was open.
 TILEID   int32          ID of the observed tile.
 SNR2FRAC float32        Fractional SNR2 accumulated on this tile during this exposure.
 AIRMASS  float32        Average airmass during this exposure.
-SEEING   float32 arcsec Average FWHM atmospheric seeing during this exposure. 
+SEEING   float32 arcsec Average FWHM atmospheric seeing during this exposure.
 TRANSP   float32        Average atmospheric transparency during this exposure.
 SKY      float32        Average sky background level during this exposure.
 ======== ======= ====== ===========
@@ -96,7 +93,6 @@ KEY     Example Value Type Comment
 ======= ============= ==== =====================
 NAXIS1  20            int  length of dimension 1
 NAXIS2  10            int  length of dimension 2
-EXTNAME TILEDATA      str  extension name
 ======= ============= ==== =====================
 
 Required Data Table Columns
@@ -112,7 +108,7 @@ SNR2FRAC float32       Total fractional SNR2 accumulated on this tile.
 NEXP     int32         Total number of exposures of this tile.
 ======== ======= ===== ===========
 
-There is one table row per tile, indexed to match `desisurvey.tiles.Tiles 
+There is one table row per tile, indexed to match `desisurvey.tiles.Tiles
 <https://desisurvey.readthedocs.io/en/latest/api.html#desisurvey.tiles.Tiles>`__.
 
 The integer ``AVAIL`` and ``PLANNED`` values are nights since the date specified
@@ -135,11 +131,11 @@ object records exposures during simulation::
 
     import surveysim.exposures
     explist = surveysim.exposures.ExposureList()
-    
+
 Its internal state after a simulation (or each night) can be saved using, for example::
 
     explist.save('exposures.fits', comment='Baseline (seed=1)')
-    
+
 This state can then later be restored using::
 
     explist = surveysim.exposures.ExposureList(restore='exposures.fits')
