+++ doc/DESI_SPECTRO_REDUX/SPECPROD/spectra-NSIDE/PIXGROUP/PIXNUM/zbest-NSIDE-PIXNUM.rst
@@ -96,7 +96,6 @@ SPECTYPE  char[6]           Spectral type
 SUBTYPE   char[6]           Spectral subtype (maybe blank)
 NCOEFF    int64
 DELTACHI2 float64           Delta(chi2) to next best fit
-BRICKNAME char[8]           Imaging brickname where this target came from
 NUMEXP    int32
 NUMTILE   int32
 ========= =========== ===== =============================================
@@ -139,8 +138,10 @@ SUBPRIORITY       float64
 REF_ID            int64
 PMRA              float32
 PMDEC             float32
+REF_EPOCH         float32
 PMRA_IVAR         float32
 PMDEC_IVAR        float32
+RELEASE           int16
 FLUX_G            float32
 FLUX_R            float32
 FLUX_Z            float32
@@ -166,6 +167,11 @@ MW_TRANSMISSION_R float32
 MW_TRANSMISSION_Z float32
 EBV               float32
 PHOTSYS           char[1]
+OBSCONDITIONS     int32
+NUMOBS_INIT       int64
+PRIORITY_INIT     int64
+NUMOBS_MORE       int32
+HPXPIXEL          int64
 FIBER             int32         Fiber ID [0-4999]
 PETAL_LOC         int32
 DEVICE_LOC        int32
@@ -173,23 +179,25 @@ LOCATION          int32
 FIBERSTATUS       int32
 OBJTYPE           char[3]
 LAMBDA_REF        float32
-DESIGN_X          float32
-DESIGN_Y          float32
-DESIGN_Q          float32
-DESIGN_S          float32
+FIBERASSIGN_X     float32
+FIBERASSIGN_Y     float32
+FA_TARGET         int64
+FA_TYPE           byte
 NUMTARGET         int16
 FIBER_RA          float64
 FIBER_DEC         float64
 FIBER_RA_IVAR     float32
 FIBER_DEC_IVAR    float32
-DELTA_X           float32
-DELTA_Y           float32
-DELTA_X_IVAR      float32
-DELTA_Y_IVAR      float32
+PLATEMAKER_X      float32
+PLATEMAKER_Y      float32
+PLATEMAKER_RA     float32
+PLATEMAKER_DEC    float32
 NUM_ITER          int32
 SPECTROID         int32
 BRICKNAME         char[8]       Brickname from target imaging
 LAMBDAREF         float64       Reference wavelength at which to align fiber
+DELTA_X           float64
+DELTA_Y           float64
 NIGHT             int32
 EXPID             int32
 TILEID            int32
