--- doc/QUICKLOOK/index.rst
@@ -1,23 +1,0 @@
-=========
-QUICKLOOK
-=========
-
-.. toctree::
-   :maxdepth: 1
-
-   $ql-checkHDUs-{camera}-{expid}.json : QA Result of matching camera and expid in the raw header <ql-checkHDUs-CAMERA-EXPID>
-   $ql-checkfibflat-{camera}-{expid}.json : QA output for examining fiber flat and test for stability and consistency <ql-checkfibflat-CAMERA-EXPID>
-   $ql-checkres-{camera}-{expid}.json : QA output for examining W sigmas and evaluates the resolution fitting solution for the arc lines <ql-checkres-CAMERA-EXPID>
-   $ql-countbins-{camera}-{expid}.json : QA output counting bins in spectrum above thresholds <ql-countbins-CAMERA-EXPID>
-   $ql-countpix-{camera}-{expid}.json : QA output counting pixels above threshold  <ql-countpix-CAMERA-EXPID>
-   $ql-getbias-{camera}-{expid}.json : QA output concerning bias in oversan <ql-getbias-CAMERA-EXPID>
-   $ql-getrms-{camera}-{expid}.json : QA output concerning RMS in preprocessed image <ql-getrms-CAMERA-EXPID>
-   $ql-integ-{camera}-{expid}.json : QA output calculating total counts for standard stars <ql-integ-CAMERA-EXPID>
-   $ql-skycont-{camera}-{expid}.json : QA output from sky continuum calculation <ql-skycont-CAMERA-EXPID>
-   $ql-skypeak-{camera}-{expid}.json : QA output calculating statistics at peak sky wavelengths <ql-skypeak-CAMERA-EXPID>
-   $ql-skyRband-{camera}-{expid}.json : QA output comparing skyflux in Rband with that of the skymonitor <ql-skyRband-CAMERA-EXPID>
-   $ql-snr-{camera}-{expid}.json : QA output for the signal to noise calculation after sky subtraction <ql-snr-CAMERA-EXPID>
-   $ql-trace-{camera}-{expid}.json : QA output calculating the average shifts in pixels in X and Y directions over all 500 fibers <ql-trace-CAMERA-EXPID>
-   $ql-xwsigma-{camera}-{expid}.json : QA output fitting X and W sigmas to preprocessed spectra <ql-xwsigma-CAMERA-EXPID>
-   $ql-mergedQA-{camera}-{expid}.json : QAs grouped by tasks CHECK_HDUs, CHECK_CCDs, CHECK_FIBERS, CHECK_SPECTRA <ql-mergedQA-CAMERA-EXPID>
-
