+++ doc/index.rst
@@ -21,7 +21,6 @@ for testing.
    $DESI_TARGET : target selection and fiber assignment <DESI_TARGET/index>
    $DESIMODEL : data used for simulating DESI <DESIMODEL/index>
    $PROTODESI : data and logs from the ProtoDESI campaign <PROTODESI/index>
-   $QUICKLOOK : data outputs form QUICKLOOK algorithms <QUICKLOOK/index>
    $DESISURVEY_OUTPUT : outputs from desisurvey and surveysim <DESISURVEY_OUTPUT/index>
    datamodel
    changes
