+++ setup.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-from __future__ import absolute_import, division, print_function
 #
 # Standard imports
 #
@@ -15,7 +14,7 @@ from setuptools import setup, find_packages
 #
 # DESI support code.
 #
-from desiutil.setup import DesiTest, DesiVersion, get_version
+import desiutil.setup as ds
 #
 # Begin setup
 #
@@ -32,7 +31,7 @@ setup_keywords['url'] = 'https://github.com/desihub/desidatamodel'
 #
 # END OF SETTINGS THAT NEED TO BE CHANGED.
 #
-setup_keywords['version'] = get_version(setup_keywords['name'])
+setup_keywords['version'] = ds.get_version(setup_keywords['name'])
 #
 # Use README.rst as long_description.
 #
@@ -50,14 +49,17 @@ if os.path.isdir('bin'):
     setup_keywords['scripts'] = [fname for fname in glob.glob(os.path.join('bin', '*'))
         if not os.path.basename(fname).endswith('.rst')]
 setup_keywords['provides'] = [setup_keywords['name']]
-setup_keywords['requires'] = ['Python (>3.5.0)']
-# setup_keywords['install_requires'] = ['Python (>3.5.0)']
+setup_keywords['python_requires'] = '>=3.5'
 setup_keywords['zip_safe'] = False
 setup_keywords['use_2to3'] = False
 setup_keywords['packages'] = find_packages('py')
-setup_keywords['package_dir'] = {'':'py'}
-setup_keywords['cmdclass'] = {'version': DesiVersion,'test': DesiTest, 'sdist': DistutilsSdist}
-setup_keywords['test_suite']='{name}.test.{name}_test_suite.{name}_test_suite'.format(**setup_keywords)
+setup_keywords['package_dir'] = {'': 'py'}
+setup_keywords['cmdclass'] = {'module_file': ds.DesiModule,
+                              'version': ds.DesiVersion,
+                              'test': ds.DesiTest,
+                              'api': ds.DesiAPI,
+                              'sdist': DistutilsSdist}
+setup_keywords['test_suite']='{name}.test.{name}_test_suite'.format(**setup_keywords)
 #
 # Autogenerate command-line scripts.
 #
